なんとなくファイルを切りだした。意味あんのかは微妙。
